requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
mysqlclient==2.2.0
pycodestyle==2.11.1
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
